refactor(admin): extract Jobs tab into pages/admin/jobs.vue
